src/ratingservice/requirements.txt
flask==1.1.2
psycopg2-binary==2.8.5
gunicorn==0.17.0